Fix CI build errors and test warnings
- Update actions/upload-artifact from v3 to v4 in all workflow jobs
- Fix signed/unsigned integer comparison warnings in test files

+++ test/test_cat.c
@@ -140,7 +140,7 @@ void test_id_array_bounds(void)
     Cat testCat;
     
     // Test all 6 bytes of ID
-    for (int i = 0; i < 6; i++) {
+    for (uint8_t i = 0; i < 6; i++) {
         testCat.id[i] = (uint8_t)i;
         TEST_ASSERT_EQUAL_UINT8(i, testCat.id[i]);
     }
